refactor(sdk): move parse_retry_after out of exceptions.py
exceptions.py was a zero-import, pure exception-definitions module before
this PR (matching errors.ts on the JS side); the Retry-After parser had
been added there directly, pulling in datetime/email.utils/math imports
that belong to the parsing algorithm, not to the exception classes.

Move it to a new http_utils.py (mirroring this SDK's existing io_utils.py
naming, and matching where the JS side already put the equivalent
parseRetryAfter helper: utils.ts, not errors.ts). exceptions.py keeps
only the `Optional` import needed for RateLimitException's own
retry_after field.

Also extract the JS-side retryAfter ternary in handleApiError into a
named const instead of computing it inline as a positional function
argument, matching the pattern already used by every comparable case in
this codebase (connectionConfig.ts, sandboxApi.ts, template/index.ts,
and envd/api.ts's own analogous `running` computation).
